OSIS-5248 removed Check2M validator and reuse unit tests on AttachFinalityEndDateValidator
